requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
discord.py==1.6.0
docutils==0.16
idna==2.10
isort==5.7.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.1
kivy-deps.sdl2==0.3.1
Kivy-examples==1.11.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.1
mccabe==0.6.1
multidict==5.1.0
mysqlclient==2.0.1
numpy==1.19.1
Pillow==7.2.0
pygame==2.0.0
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.15.0
pywin32==228
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
